Add files via upload
For connecting database:

Make a database.
Add a user to that database with all permissions possible.

Go back to the project and open the applications.properties file. Change the username and password to the database's login credentials.

To fix errors:
In the applications.properties file, there is a typo. At one point 'econjava' is written instead of 'ecommjava'   
To check if the database has been connected:
You will find some new tables in the database after you run the code. At this point, the database has been connected but there is still a bug in the program. Somewhere a wrong function has been used. 'Execute' instead of 'executeQuery'